+++ changelog.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v0.5.0
+- Added workflow to validate release branch.
 - Added method to do query with more complex conditions.
 - Added method to get count records in a entity.
 - Added support query with like operator.